clean up to individual files
